image effects

+++ frontend/src/Components/Profile.js
@@ -41,7 +41,6 @@ export default class Profile extends Component {
 
   
     editProfileHandler = (e) => {
-        console.log(this.props)
         const loggedinuser = this.state;
         this.setState({editprofile : 1})
         this.props.history.push(`/editprofile/${this.state.email}`)
@@ -49,16 +48,17 @@ export default class Profile extends Component {
     }
 
     deleteAccountHandler = () => {
-        console.log('deleteing')
-        axios.delete('http://localhost:3001/deleteaccount/' + sessionStorage.useremail)
-        .then(res => {
-            sessionStorage.clear();
-            this.props.history.push('/')
-            toast.success("Account deleted")
-        })
-        .catch(err => {
-            console.log(err)
-        })
+        if (window.confirm("Are you sure you want to leave?")) {
+            axios.delete('http://localhost:3001/deleteaccount/' + sessionStorage.useremail)
+            .then(res => {
+                sessionStorage.clear();
+                this.props.history.push('/')
+                toast.success("Account deleted")
+            })
+            .catch(err => {
+                console.log(err)
+            })
+        }
     }
    
     render() {
